reuse count util
