+++ packages/dataviews/src/components/dataform-layouts/card/index.tsx
@@ -26,7 +26,6 @@ import { getFormFieldLayout } from '..';
 import DataFormContext from '../../dataform-context';
 import type {
 	FieldLayoutProps,
-	FieldValidity,
 	NormalizedCardLayout,
 	NormalizedField,
 	NormalizedForm,
@@ -36,33 +35,7 @@ import { DataFormLayout } from '../data-form-layout';
 import { DEFAULT_LAYOUT } from '../normalize-form';
 import { getSummaryFields } from '../get-summary-fields';
 import useReportValidity from '../../../hooks/use-report-validity';
-
-function countInvalidFields( validity: FieldValidity | undefined ): number {
-	if ( ! validity ) {
-		return 0;
-	}
-
-	let count = 0;
-	const validityRules = Object.keys( validity ).filter(
-		( key ) => key !== 'children'
-	);
-
-	for ( const key of validityRules ) {
-		const rule = validity[ key as keyof Omit< FieldValidity, 'children' > ];
-		if ( rule?.type === 'invalid' ) {
-			count++;
-		}
-	}
-
-	// Count children recursively
-	if ( validity.children ) {
-		for ( const childValidity of Object.values( validity.children ) ) {
-			count += countInvalidFields( childValidity );
-		}
-	}
-
-	return count;
-}
+import countInvalidFields from '../count-invalid-fields';
 
 const NonCollapsibleCardHeader = ( {
 	children,
